Switch to 32 bit values for most places

+++ src/level.rs
@@ -107,12 +107,12 @@ pub(crate) struct ParsedInnerLevel<'a> {
 }
 
 #[derive(Default, Resource)]
-pub(crate) struct Groups(pub(crate) HashMap<u64, Group>);
+pub(crate) struct Groups(pub(crate) HashMap<u32, Group>);
 
 pub(crate) struct Group {
     pub(crate) entities: Vec<Entity>,
     pub(crate) activated: bool,
-    pub(crate) opacity: f64,
+    pub(crate) opacity: f32,
 }
 
 impl Default for Group {
@@ -131,8 +131,8 @@ impl<'a> ParsedInnerLevel<'a> {
         commands: &mut Commands,
         low_detail: bool,
     ) -> Result<(), anyhow::Error> {
-        let mut colors: HashMap<u64, ColorChannel> = HashMap::new();
-        let mut groups: HashMap<u64, Group> =
+        let mut colors: HashMap<u32, ColorChannel> = HashMap::new();
+        let mut groups: HashMap<u32, Group> =
             HashMap::with_capacity((self.objects.len() / 500).min(500));
         if let Some(colors_string) = self.start_object.get(b"kS38".as_ref()) {
             let parsed_colors: Vec<&[u8]> = de::from_slice(colors_string, b'|')?;
@@ -170,7 +170,7 @@ impl<'a> ParsedInnerLevel<'a> {
                     continue;
                 }
             }
-            let parsed_groups: Vec<u64> =
+            let parsed_groups: Vec<u32> =
                 if let Some(group_string) = object_data.get(b"57".as_ref()) {
                     de::from_slice(group_string, b'.')?
                 } else {
